:recycle: invoke offline mode test separately
